Fix the self.actionSetFontSizeAndIndent code
At first when I tested the new widget, I couldn't open it. I was sure that there was no error in the code... turns out I need to change the "widget" variable from "local" to "self".

And also, QPlainTextEdit has no "setFontPointSize()" function, so I had to work around this: change the size with the "setFont()" function, using a QFont object.